add "share this page" link
This adds a "share this page" link to the bottom of the home page. With
javascript enabled, the link makes a new section appear at the bottom, with
links to share on many popular sites. Without javascript (so everything is on
one page) the "share this page" link scrolls you down to the bottom where the
share links are.

No icons were used. This is easy, fast, and there's no potential licensing or
trademark issues.

[gp: re-arrage code a bit, + make share part a bit less obtrusive visually]